comment,board 세션 추가 작업

+++ Back/src/main/java/com/spring/back/service/BoardServiceImpl.java
@@ -59,9 +59,8 @@ public class BoardServiceImpl implements BoardService {
 		Board board = BoardDTO.boardDtotoEntity(boardDTO);
 
 		// board에 user 직접 삽입하기
-		User user = userRepo.findByUserId(session.getUser().getUserId());
-
-		board.updateUser(user);
+		board.updateUser(userRepo.findByUserId(session.getUser().getUserId()));
+		
 		return boardRepo.save(board).getBoardNo();
 	}
 
@@ -71,7 +70,7 @@ public class BoardServiceImpl implements BoardService {
 	@Override
 	public List<BoardDTO> findBoardsByPage(PageRequest pageRequest) {
 		return boardRepo.findAll(pageRequest).stream()
-				.map(board -> Board.boardEntityToDTO(board))
+				.map(board -> Board.FalseEntityToDTO(board))
 				.collect(Collectors.toList());
 	}
 
@@ -86,10 +85,10 @@ public class BoardServiceImpl implements BoardService {
 		board.updateViewCount(board.getViewCount() + 1);
 	
 		if (session.getUser().getUserId().equals(board.getUser().getUserId())) {
-			BoardDTO boardDTO = Board.StatusboardEntityToDTO(board);
+			BoardDTO boardDTO = Board.TrueboardEntityToDTO(board);
 			return boardDTO;
 		} else {
-			BoardDTO boardDTO = Board.boardEntityToDTO(board);
+			BoardDTO boardDTO = Board.FalseEntityToDTO(board);
 			return boardDTO;
 		}
 		
@@ -101,7 +100,7 @@ public class BoardServiceImpl implements BoardService {
 	@Transactional
 	public BoardDTO findBoardByBoardNo(Long boardNo) {
 		Board board = boardRepo.findById(boardNo).orElseThrow(NoSuchElementException::new);
-		BoardDTO boardDTO = Board.boardEntityToDTO(board);
+		BoardDTO boardDTO = Board.FalseEntityToDTO(board);
 	
 			return boardDTO;
 			
@@ -115,7 +114,7 @@ public class BoardServiceImpl implements BoardService {
 	public List<BoardDTO> findRecoBoard(Long boardNo) {
 		// 추천 게시글 불러오기
 		List<Board> recoBoard = boardRepo.findRecommendedBoardByBoardNo(boardNo);
-		List<BoardDTO> boardDTOs = recoBoard.stream().map(board -> Board.boardEntityToDTO(board)).collect(Collectors.toList());
+		List<BoardDTO> boardDTOs = recoBoard.stream().map(board -> Board.FalseEntityToDTO(board)).collect(Collectors.toList());
 		return boardDTOs;
 	}
 	
@@ -145,15 +144,19 @@ public class BoardServiceImpl implements BoardService {
 	// [게시글 수정]
 	@Override
 	@Transactional
-	public BoardDTO updateBoard(BoardDTO newboardDTO, List<MultipartFile> files) {
+	public boolean updateBoard(SessionDTO sessionDTO, BoardDTO newboardDTO, List<MultipartFile> files) {
 		Board board = boardRepo.findById(newboardDTO.getBoardNo()).orElseThrow(NoSuchElementException::new);
-		board.updateBoard(newboardDTO);
-		BoardDTO boardDTO = Board.boardEntityToDTO(board);
-		// 기존 File 삭제
-		fileService.deleteFileBoardNo(boardDTO.getBoardNo());
-		// 새로운 File 추가
-		fileService.uploadFile(boardDTO.getBoardNo(), files);
-		return boardDTO;
+		Session session = sessionRepo.findBySessionId(sessionDTO.getSessionId());
+		if(board.getUser().equals(session.getUser())) {
+			board.updateBoard(newboardDTO);
+			BoardDTO boardDTO = Board.FalseEntityToDTO(board);
+			// 기존 File 삭제
+			fileService.deleteFileBoardNo(boardDTO.getBoardNo());
+			// 새로운 File 추가
+			fileService.uploadFile(boardDTO.getBoardNo(), files);
+			return true;
+		}
+		return false;
 	}
 	
 	// [좋아요 수 +1]
@@ -162,7 +165,7 @@ public class BoardServiceImpl implements BoardService {
 	public BoardDTO updateLikeCount(Long boardNo) {
 		Board board = boardRepo.findById(boardNo).orElseThrow(NoSuchElementException::new);
 		board.updateLikeCount(board.getLikeCount()+1);
-		BoardDTO boardDTO = Board.boardEntityToDTO(board);
+		BoardDTO boardDTO = Board.FalseEntityToDTO(board);
 		return boardDTO;
 	}
 
@@ -171,11 +174,11 @@ public class BoardServiceImpl implements BoardService {
 	// [게시글 삭제]
 	@Override
 	@Transactional
-	public boolean deleteBoard(BoardDTO boardDTO) {
+	public boolean deleteBoard(SessionDTO sessionDTO, BoardDTO boardDTO) {
 		Board deleteBoard = boardRepo.findById(boardDTO.getBoardNo()).orElseThrow(NoSuchElementException::new);
-
+		Session session = sessionRepo.findBySessionId(sessionDTO.getSessionId());
 		// 삭제 요청한 userId와 요청한 게시글의 userId가 같을 경우에만 삭제 진행
-		if (deleteBoard.getUser().getUserId().equals(boardDTO.getUserId())) {
+		if (deleteBoard.getUser().equals(session.getUser())) {
 			// file과 comment먼저 삭제 후 게시글 삭제 진행
 			commentService.deleteAllComment(boardDTO.getBoardNo());
 			fileService.deleteFileBoardNo(boardDTO.getBoardNo());
